Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "ruff>=0.14.4",
    "uvicorn>=0.38.0",
]
